Fixed #309 and updated the menu command
Fixed #309 (renamepet was not an option), and made each menu only able to be opened if the category/feature is enabled. Checking if the category or feature is enabled will prevent issues when attempting to open menus that don't have requirements met (e.g. LibsDisguises not installed, category disabled, treasure chests disabled, etc).